rename safe-connect to walletConnec
