--- public/safe-connect/manifest.json
@@ -1,6 +1,0 @@
-{
-  "short_name": "Safe WalletConnect",
-  "name": "Safe Connect",
-  "description": "Allows your Gnosis Safe Multisig to connect to dapps via WalletConnect.",
-  "iconPath": "safe-connect.jpg"
-}
