+++ public/walletConnect/manifest.json
@@ -1,0 +1,6 @@
+{
+  "short_name": "WalletConnect",
+  "name": "WalletConnect",
+  "description": "Allows your Gnosis Safe Multisig to connect to dapps via WalletConnect.",
+  "iconPath": "walletConnect.jpg"
+}
