+++ src/apps/WalletConnect/App.tsx
