+++ src/apps/WalletConnect/components/WCClientInfo.tsx
