+++ src/apps/WalletConnect/components/layout.tsx
