+++ src/apps/WalletConnect/hooks/useWalletConnect.tsx
