+++ src/apps/WalletConnect/index.tsx
