+++ src/apps/WalletConnect/utils.ts
